LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-option-element/option-text-backslash-expected.txt


FAIL Test for the backslash sign in option.text assert_equals: expected "\\" but got "¥"
